tried replacing float -> double

+++ Firmware/shared/include/utils.hpp
@@ -1,21 +1,19 @@
 #pragma once
 
-#include <Arduino.h>
-
 struct Vector2D
 {
-    float x, y;
+    double x, y;
 
     Vector2D() {}
-    Vector2D(float _x, float _y) : x(_x), y(_y) {}
+    Vector2D(double _x, double _y) : x(_x), y(_y) {}
 
-    static Vector2D fromPolar(float angle, float length);
-    float length();
-    float angle();
+    static Vector2D fromPolar(double angle, double length);
+    double length();
+    double angle();
     Vector2D operator+(const Vector2D &other);
     Vector2D operator-(const Vector2D &other);
-    float operator*(const Vector2D &other);
-    Vector2D operator*(const float scale);
+    double operator*(const Vector2D &other);
+    Vector2D operator*(const double scale);
 };
 
-float determinant(Vector2D first, Vector2D second);
+double determinant(Vector2D first, Vector2D second);
