+++ Firmware/shared/lib/physics/collider.cpp
@@ -57,7 +57,7 @@ bool Collider::contains(Vector2D point)
 bool Collider::getEnteringEdge(Vector2D handlePosition, Vector2D objectPosition, Edge* enteringEdge)
 {
     // will contain result
-    auto minDist = 0.0f;
+    auto minDist = 0.0;
     auto foundAny = false;
     // loop vars
     auto edgeCount = m_points.size();
@@ -65,7 +65,7 @@ bool Collider::getEnteringEdge(Vector2D handlePosition, Vector2D objectPosition,
     // pre-allocate
     Vector2D first, second, intersection;
     bool intersects;
-    float scale, dist;
+    double scale, dist;
 
     for(auto i = 0; i < edgeCount; ++i)
     {
@@ -121,5 +121,5 @@ Vector2D Collider::getClosestOutsidePoint(Edge edge, Vector2D handlePosition)
         false);
     
     auto collisionVec = intersection - handlePosition;
-    return handlePosition + collisionVec * 1.1f;
+    return handlePosition + collisionVec * 1.1;
 }
